ggtime: Grammar of Graphics and Plot Helpers for Time Series Visualization

Extends the capabilities of 'ggplot2' by providing grammatical elements and plot helpers designed for visualizing temporal patterns. The package implements a grammar of temporal graphics, which leverages calendar structures to highlight changes over time. The package also provides plot helper functions to quickly produce commonly used time series graphics, including time plots, season plots, and seasonal sub-series plots.

Author

Maintainer: Mitchell O'Hara-Wild mail@mitchelloharawild.com (ORCID)

Authors:

Cynthia A. Huang (ORCID)

Matthew Kay (ORCID)

Rob Hyndman (ORCID)