Calendar coordinates

The calendar coordinate system arranges time series data into a calendar-like layout, making it easier to see fine-grained temporal patterns over a long time span. It has similar semantics as the looped coordinate system (coord_loop()), however instead of overlaying looped data the calendar coordinate space arranges each loop into rows and columns like a calendar.

coord_calendar(
  rows = waiver(),
  time_rows = waiver(),
  cols = waiver(),
  time_cols = waiver(),
  time = "x",
  ljust = 0.5,
  xlim = NULL,
  ylim = NULL,
  expand = FALSE,
  default = FALSE,
  clip = "on",
  clip_rows = "on",
  coord = coord_cartesian()
)

Arguments

rows

Layout the time scale into calendar rows, one of:

NULL or waiver() for no rows (the default)

A mixtime vector giving time points at which the time axis should layout into rows

A function that takes the limits as input and returns row layout points as output

time_rows

A duration giving the distance between calendar rows like "1 weeks", or "1 month". If both rows and time_rows are specified, time_rows wins.

cols, time_cols

Not yet supported.

time

A string specifying which aesthetic contains the time variable that should be looped over. Default is "x".

ljust

Loop justification, a number between 0 and 1 indicating where the lines between looped positions are drawn (0 indicates left, 1 indicates right, 0.5 indicates center).

xlim, ylim

Limits for the x and y axes. NULL means use the default limits.

expand

Logical indicating whether to expand the coordinate limits. Default is FALSE.

default

Logical indicating whether this is the default coordinate system. Default is FALSE.

clip

Should drawing be clipped to the extent of the plot panel? A setting of "on" (the default) means yes, and a setting of "off" means no.

clip_rows

Should the drawing of each loop of the timescale be clipped to the breaks defined by time_rows and ljust? A setting of "on" (the default) means yes, and a setting of "off" means no.

coord

The underlying coordinate system to use. Default is coord_cartesian().

Details

This coordinate system is particularly useful for visualizing long time spans with events that occur over short intervals (such as holidays).

It works by:

Dividing the time axis into segments based on the specified row (and column) periods

Translating each panel into the rows and columns of a calendar layout

The coordinate system requires R version 4.2.0 or higher due to its use of usage of clipping paths.

Examples

# Visualise several months of Melbourne pedestrian counts across weekly rows
tsibble::pedestrian |>
  dplyr::filter(Date < "2015-04-01") |>
  ggplot(aes(x = Date_Time, y = Count, color = Sensor)) +
  geom_line() +
  coord_calendar(time_rows = "1 week")
#> Error in ggplot(dplyr::filter(tsibble::pedestrian, Date < "2015-04-01"),     aes(x = Date_Time, y = Count, color = Sensor)): could not find function "ggplot"